+++ lib/src/rest/request_builder.dart
@@ -37,6 +37,40 @@ class PostBuilder {
   }
 }
 
+class PatchBuilder {
+  String _url;
+
+  PatchBuilder(this._url);
+
+  Map<String, String>? _header = <String, String>{};
+  Map<String, dynamic> _body = <String, dynamic>{};
+
+  void setHeader(HeaderBuilder headerBuilder) {
+    _header = headerBuilder.get();
+  }
+
+  void setBody(Map<String, dynamic> body) {
+    this._body = body;
+  }
+
+  void setQueryParams(Map<String, String>? queryParams) {
+    if (queryParams == null || queryParams.isEmpty) {
+      return;
+    }
+    _url = _url + '?';
+
+    queryParams.forEach((key, value) {
+      _url = _url + key + '=' + value + '&';
+    });
+    _url = _url.substring(0, _url.length - 1); //remove trailing &
+  }
+
+  Future<http.Response> patch() async {
+    return http.patch(Uri.parse(_url),
+        headers: _header, body: jsonEncode(_body));
+  }
+}
+
 class PutBuilder {
   String _url;
 
@@ -107,16 +141,37 @@ class GetBuilder {
   }
 }
 
-/*
-class HeaderBuilder {
-  static Map<String, String> get(AppContext appContext) {
-    Map<String, String> _header = new Map<String, String>();
-    _header['Content-Type'] = 'application/json';
+class DeleteBuilder {
+  String _interpretedURL;
 
-    if (appContext.hasToken) {
-      _header['authorization'] = 'Bearer ' + appContext.getToken.token;
+  DeleteBuilder(String url) : _interpretedURL = url;
+
+  var _header = new Map<String, String>();
+
+  void setHeader(HeaderBuilder headerBuilder) {
+    _header = headerBuilder.get();
+  }
+
+  void setURLParam(String key, String value) {
+    _interpretedURL = _interpretedURL.replaceAll("{" + key + "}", value);
+  }
+
+  //Must be called after above
+  void setQueryParams(Map<String, String>? queryParams) {
+    if (queryParams == null || queryParams.isEmpty) {
+      return;
     }
-    return _header;
+    _interpretedURL = _interpretedURL + '?';
+
+    queryParams.forEach((key, value) {
+      _interpretedURL = _interpretedURL + key + '=' + value + '&';
+    });
+    _interpretedURL = _interpretedURL.substring(
+        0, _interpretedURL.length - 1); //remove trailing &
+  }
+
+  Future<http.Response> delete() async {
+    Uri _uri = Uri.parse(_interpretedURL);
+    return http.delete(_uri, headers: _header);
   }
 }
-*/