+++ lib/src/rest/url_factory.dart
@@ -7,39 +7,55 @@ class URLFactory {
   Map<dynamic, String> getPaths = <dynamic, String>{};
   Map<dynamic, String> postPaths = <dynamic, String>{};
   Map<dynamic, String> putPaths = <dynamic, String>{};
+  Map<dynamic, String> deletePaths = <dynamic, String>{};
+  Map<dynamic, String> patchPaths = <dynamic, String>{};
 
-  void addURL<T>(Protocol protocol, String path) {
-    switch (protocol) {
-      case Protocol.get:
+  void addPath<T>(Method method, String path) {
+    switch (method) {
+      case Method.delete:
+        deletePaths[T] = path;
+        break;
+      case Method.get:
         getPaths[T] = path;
         break;
-      case Protocol.post:
+      case Method.post:
         postPaths[T] = path;
         break;
-      case Protocol.put:
+      case Method.put:
         putPaths[T] = path;
         break;
-      case Protocol.all:
+      case Method.patch:
+        patchPaths[T] = path;
+        break;
+      case Method.all:
+        deletePaths[T] = path;
         getPaths[T] = path;
         postPaths[T] = path;
         putPaths[T] = path;
+        patchPaths[T] = path;
         break;
     }
   }
 
-  String getURL<T>(Protocol protocol) {
+  String getURL<T>(Method method) {
     String? path;
-    switch (protocol) {
-      case Protocol.get:
+    switch (method) {
+      case Method.delete:
+        path = deletePaths[T];
+        break;
+      case Method.get:
         path = getPaths[T];
         break;
-      case Protocol.post:
+      case Method.post:
         path = postPaths[T];
         break;
-      case Protocol.put:
+      case Method.put:
         path = putPaths[T];
         break;
-      case Protocol.all:
+      case Method.patch:
+        path = patchPaths[T];
+        break;
+      case Method.all:
         path = getPaths[T];
         break;
     }